# fix boilerplate check for time-to-k8s git submodule